# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ControlHub SaaS Empresarial API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /companies/{companyId}/employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List employees
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: area
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create employee
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /companies/{companyId}/employees/{employeeId}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: employeeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeDetail"
    put:
      operationId: updateEmployee
      tags: [employees]
      summary: Update employee
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: employeeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmployeeBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete employee
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: employeeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # INVOICES
  /companies/{companyId}/invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List invoices
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: supplierId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create invoice
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /companies/{companyId}/invoices/{invoiceId}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice detail
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: invoiceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    put:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update invoice
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: invoiceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInvoiceBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      summary: Delete invoice
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: invoiceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # SUPPLIERS
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Employee:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        documentId:
          type: string
        position:
          type: string
        area:
          type: string
        site:
          type: string
          nullable: true
        contractType:
          type: string
        status:
          type: string
          enum: [active, inactive, on_leave, suspended]
        startDate:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        avatar:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, companyId, firstName, lastName, documentId, position, area, contractType, status, startDate, createdAt]
    EmployeeDetail:
      allOf:
        - $ref: "#/components/schemas/Employee"
        - type: object
          properties:
            address:
              type: string
              nullable: true
            emergencyContact:
              type: string
              nullable: true
            emergencyPhone:
              type: string
              nullable: true
            notes:
              type: string
              nullable: true
            attendanceSummary:
              type: object
              properties:
                presentDays:
                  type: integer
                absentDays:
                  type: integer
                lateDays:
                  type: integer
                hoursWorked:
                  type: number
              required: [presentDays, absentDays, lateDays, hoursWorked]
            recentAnnouncements:
              type: array
              items:
                $ref: "#/components/schemas/Announcement"
            documentCount:
              type: integer
          required: [attendanceSummary, recentAnnouncements, documentCount]
    CreateEmployeeBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        documentId:
          type: string
        position:
          type: string
        area:
          type: string
        site:
          type: string
        contractType:
          type: string
          enum: [indefinite, fixed_term, part_time, intern, contractor]
        status:
          type: string
          enum: [active, inactive, on_leave, suspended]
        startDate:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        emergencyContact:
          type: string
        emergencyPhone:
          type: string
        notes:
          type: string
      required: [firstName, lastName, documentId, position, area, contractType, status, startDate]
    UpdateEmployeeBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        position:
          type: string
        area:
          type: string
        site:
          type: string
        contractType:
          type: string
        status:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        emergencyContact:
          type: string
        emergencyPhone:
          type: string
        notes:
          type: string
    Invoice:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        supplierId:
          type: integer
        supplierName:
          type: string
        invoiceNumber:
          type: string
        issueDate:
          type: string
        dueDate:
          type: string
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [pending, paid, overdue, cancelled, observed, scheduled]
        category:
          type: string
          nullable: true
        costCenter:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        fileUrl:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, companyId, supplierId, supplierName, invoiceNumber, issueDate, dueDate, amount, currency, status, createdAt]
    CreateInvoiceBody:
      type: object
      properties:
        supplierId:
          type: integer
        invoiceNumber:
          type: string
        issueDate:
          type: string
        dueDate:
          type: string
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [pending, paid, overdue, cancelled, observed, scheduled]
        category:
          type: string
        costCenter:
          type: string
        notes:
          type: string
        fileUrl:
          type: string
      required: [supplierId, invoiceNumber, issueDate, dueDate, amount, currency, status]
    UpdateInvoiceBody:
      type: object
      properties:
        supplierId:
          type: integer
        invoiceNumber:
          type: string
        issueDate:
          type: string
        dueDate:
          type: string
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [pending, paid, overdue, cancelled, observed, scheduled]
        category:
          type: string
        costCenter:
          type: string
        notes:
          type: string
        fileUrl:
          type: string
...
    Announcement:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        title:
          type: string
        content:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
        targetType:
          type: string
          enum: [all, area, site, individual]
        targetValue:
          type: string
          nullable: true
        publishedBy:
          type: string
        readCount:
          type: integer
        isRead:
          type: boolean
        createdAt:
          type: string
      required: [id, companyId, title, content, priority, targetType, publishedBy, readCount, isRead, createdAt]
...