# lib/api-zod/src/generated/types/createInvoiceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ControlHub SaaS Empresarial API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateInvoiceBodyStatus } from "./createInvoiceBodyStatus";

export interface CreateInvoiceBody {
  supplierId: number;
  invoiceNumber: string;
  issueDate: string;
  dueDate: string;
  amount: number;
  currency: string;
  status: CreateInvoiceBodyStatus;
  category?: string;
  costCenter?: string;
  notes?: string;
  fileUrl?: string;
}